refactor(core): move artwork pipeline to core (thumbnails + metadata cache)
First slice of step 6 — the two pure helpers under
`crates/app/src/{thumbnails,metadata_artwork}.rs` had zero Tauri /
audio coupling, so they migrate straight into
`waveflow_core::artwork::{thumbnails, metadata}` with no shape
change.

What moves:

- `artwork::thumbnails` — the SIMD-resize pipeline that produces the
  `<hash>_1x.jpg` / `<hash>_2x.jpg` variants alongside the full-size
  cover, plus the worker-thread spawn helper used by the scan and
  Deezer download paths.
- `artwork::metadata` — the blake3-keyed shared cache used by the
  Deezer / Last.fm enrichers (`download_and_cache`, `path_for_hash`,
  `existing_path`). The internal `spawn_thumbnail_job` call is
  rewritten to point at `crate::artwork::thumbnails` instead of
  `crate::thumbnails`.

Desktop wiring: `crates/app/src/{thumbnails,metadata_artwork}.rs`
become one-line re-export stubs (`pub use waveflow_core::artwork::*`)
so the 13 existing `crate::thumbnails::*` / `crate::metadata_artwork::*`
call sites keep resolving without churn — same pattern as the
`commands/*` re-exports established in step 2.

`crates/core/Cargo.toml` picks up the deps the pipeline carries
(`image` with the same lossy + lossless feature set the app uses,
`fast_image_resize`, `blake3`, `anyhow`) plus `tracing` for the
`warn!` paths that already had an app-side subscriber configured.

`cargo check --workspace --all-targets` clean; `cargo test
--workspace` 111 passed / 0 failed.

Refs #128.
